feat: push fuzzing test code into `nargo` package
