+++ tooling/nargo/Cargo.toml
@@ -13,6 +13,7 @@ license.workspace = true
 acvm.workspace = true
 fm.workspace = true
 noirc_abi.workspace = true
+noir_fuzzer.workspace = true
 noirc_driver.workspace = true
 noirc_errors.workspace = true
 noirc_frontend.workspace = true
@@ -24,6 +25,7 @@ rayon = "1.8.0"
 jsonrpc.workspace = true
 rand.workspace = true
 serde.workspace = true
+proptest.workspace = true
 
 [dev-dependencies]
 # TODO: This dependency is used to generate unit tests for `get_all_paths_in_dir`
