+++ tooling/nargo_cli/src/cli/test_cmd.rs
@@ -1,25 +1,16 @@
 use std::io::Write;
 
-use acvm::{
-    acir::{
-        circuit::Program,
-        native_types::{WitnessMap, WitnessStack},
-    },
-    BlackBoxFunctionSolver, FieldElement,
-};
+use acvm::{BlackBoxFunctionSolver, FieldElement};
 use bn254_blackbox_solver::Bn254BlackBoxSolver;
 use clap::Args;
 use fm::FileManager;
 use nargo::{
-    insert_all_files_for_workspace_into_file_manager,
-    ops::{execute_program, DefaultForeignCallExecutor, TestStatus},
-    package::Package,
-    parse_all, prepare_package,
+    insert_all_files_for_workspace_into_file_manager, ops::TestStatus, package::Package, parse_all,
+    prepare_package,
 };
 use nargo_toml::{get_package_manifest, resolve_workspace_from_toml, PackageSelection};
 use noirc_driver::{
-    check_crate, compile_no_check, file_manager_with_stdlib, CompileOptions,
-    NOIR_ARTIFACT_VERSION_STRING,
+    check_crate, file_manager_with_stdlib, CompileOptions, NOIR_ARTIFACT_VERSION_STRING,
 };
 use noirc_frontend::{
     graph::CrateName,
@@ -188,58 +179,14 @@ fn run_test<S: BlackBoxFunctionSolver<FieldElement> + Default>(
 
     let blackbox_solver = S::default();
 
-    let test_function_has_no_arguments = context
-        .def_interner
-        .function_meta(&test_function.get_id())
-        .function_signature()
-        .0
-        .is_empty();
-
-    if test_function_has_no_arguments {
-        nargo::ops::run_test(
-            &blackbox_solver,
-            &mut context,
-            test_function,
-            show_output,
-            foreign_call_resolver_url,
-            compile_options,
-        )
-    } else {
-        use noir_fuzzer::FuzzedExecutor;
-        use proptest::test_runner::TestRunner;
-
-        let compiled_program =
-            compile_no_check(&mut context, compile_options, test_function.get_id(), None, false);
-        match compiled_program {
-            Ok(compiled_program) => {
-                let runner = TestRunner::default();
-
-                let executor = |program: &Program<FieldElement>,
-                                initial_witness: WitnessMap<FieldElement>|
-                 -> Result<WitnessStack<FieldElement>, String> {
-                    execute_program(
-                        program,
-                        initial_witness,
-                        &blackbox_solver,
-                        &mut DefaultForeignCallExecutor::<FieldElement>::new(false, None),
-                    )
-                    .map_err(|err| err.to_string())
-                };
-                let fuzzer = FuzzedExecutor::new(compiled_program.into(), executor, runner);
-
-                let result = fuzzer.fuzz();
-                if result.success {
-                    TestStatus::Pass
-                } else {
-                    TestStatus::Fail {
-                        message: result.reason.unwrap_or_default(),
-                        error_diagnostic: None,
-                    }
-                }
-            }
-            Err(err) => TestStatus::CompileError(err.into()),
-        }
-    }
+    nargo::ops::run_test(
+        &blackbox_solver,
+        &mut context,
+        test_function,
+        show_output,
+        foreign_call_resolver_url,
+        compile_options,
+    )
 }
 
 fn get_tests_in_package(
