+++ tooling/nargo_cli/tests/stdlib-tests.rs
@@ -1,12 +1,8 @@
 use std::io::Write;
 use std::{collections::BTreeMap, path::PathBuf};
 
-use acvm::acir::circuit::Program;
-use acvm::acir::native_types::{WitnessMap, WitnessStack};
-use acvm::FieldElement;
 use fm::FileManager;
-use nargo::ops::{execute_program, DefaultForeignCallExecutor};
-use noirc_driver::{check_crate, compile_no_check, file_manager_with_stdlib, CompileOptions};
+use noirc_driver::{check_crate, file_manager_with_stdlib, CompileOptions};
 use noirc_frontend::hir::FunctionNameMatch;
 
 use nargo::{
@@ -51,61 +47,14 @@ fn run_stdlib_tests() {
     let test_report: Vec<(String, TestStatus)> = test_functions
         .into_iter()
         .map(|(test_name, test_function)| {
-            let test_function_has_no_arguments = context
-                .def_interner
-                .function_meta(&test_function.get_id())
-                .function_signature()
-                .0
-                .is_empty();
-
-            let status = if test_function_has_no_arguments {
-                run_test(
-                    &bn254_blackbox_solver::Bn254BlackBoxSolver,
-                    &mut context,
-                    &test_function,
-                    false,
-                    None,
-                    &CompileOptions::default(),
-                )
-            } else {
-                use noir_fuzzer::FuzzedExecutor;
-                use proptest::test_runner::TestRunner;
-
-                let compiled_program = compile_no_check(
-                    &mut context,
-                    &CompileOptions::default(),
-                    test_function.get_id(),
-                    None,
-                    false,
-                );
-                match compiled_program {
-                    Ok(compiled_program) => {
-                        let runner = TestRunner::default();
-
-                        let executor = |program: &Program<FieldElement>, initial_witness: WitnessMap<FieldElement>| -> Result<WitnessStack<FieldElement>, String> {
-                            execute_program(
-                                program,
-                                initial_witness,
+            let status = run_test(
                 &bn254_blackbox_solver::Bn254BlackBoxSolver,
-                                &mut DefaultForeignCallExecutor::<FieldElement>::new(false, None),
-                            )
-                            .map_err(|err| err.to_string())
-                        };
-                        let fuzzer = FuzzedExecutor::new(compiled_program.into(), executor, runner);
-
-                        let result = fuzzer.fuzz();
-                        if result.success {
-                            TestStatus::Pass
-                        } else {
-                            TestStatus::Fail {
-                                message: result.reason.unwrap_or_default(),
-                                error_diagnostic: None,
-                            }
-                        }
-                    }
-                    Err(err) => TestStatus::CompileError(err.into()),
-                }
-            };
+                &mut context,
+                &test_function,
+                false,
+                None,
+                &CompileOptions::default(),
+            );
             (test_name, status)
         })
         .collect();
